Estimate Propensity Score

This function estimates the propensity score using a RandomForest model.

estimate_propensity(treatment, X)

Arguments

treatment

Treatment to be studied.

X

Complete sample of the x-variables to estimate the propensity score for all the observations in the dataframe.

Value

The propensity scores for each row in our dataset.

Dependencies

randomForest

dplyr

Examples

 estimate_propensity(
   treatment = df$treatment,
   X = df[[ml_variables]]
   )
#> Loading required package: randomForest
#> Warning: package ‘randomForest’ was built under R version 4.1.3
#> randomForest 4.7-1.1
#> Type rfNews() to see new features/changes/bug fixes.
#> 
#> Attaching package: ‘randomForest’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     combine
#> [1] "Training the Random Forest Model."
#> Error in nrow(df): object 'ml_variables' not found